+++ Render3D.h
@@ -71,7 +71,7 @@ struct alignas(16) Transform3D {
         vec4(0.f, 0.f, 1.f, 0.f),
         vec4(0.f, 0.f, 0.f, 1.f)}}) { }
     Transform3D inverted() const;
-    Transform3D fit(int w, int h, int x1, int x2, int y1, int y2, int z1, int z2) const;
+    Transform3D fit(int w, int h, float x1, float x2, float y1, float y2, float z1, float z2) const;
     static Transform3D translator(float dx, float dy, float dz);
     static Transform3D scaler(float x, float y, float z);
     static Transform3D rotatorX(float theta);
